int32 everywhere: block_counts, block_prefix, kernel loads

+++ train.py
@@ -17,7 +17,7 @@ def _compute_block_counts_kernel(packed_mask, block_counts,
     x = (x & 0x55) + ((x >> 1) & 0x55)
     x = (x & 0x33) + ((x >> 2) & 0x33)
     x = (x & 0x0F) + ((x >> 4) & 0x0F)
-    total = tl.sum(x, 0)
+    total = tl.sum(x, 0).to(tl.int32)
     tl.store(block_counts + pid, total)
 
 
@@ -34,7 +34,7 @@ def _reconstruct_packed_kernel(vals, packed_mask, block_prefix, out,
     m = ((byte >> bit_idx) & 1).to(tl.int32)
 
     local_idx = tl.cumsum(m, 0) - 1
-    base = tl.load(block_prefix + pid)
+    base = tl.load(block_prefix + pid).to(tl.int32)
 
     v = tl.load(vals + base + local_idx, mask=(offs < N) & (m == 1), other=0.0)
     tl.store(out + offs, tl.where(m == 1, v, 0.0), mask=offs < N)
@@ -52,13 +52,13 @@ def reconstruct_bitmask(vals, packed_mask, shape, block=8192):
     N = shape[0] * shape[1]
     n_blocks = triton.cdiv(N, block)
 
-    block_counts = torch.empty(n_blocks, device=vals.device, dtype=torch.int64)
+    block_counts = torch.empty(n_blocks, device=vals.device, dtype=torch.int32)
     _compute_block_counts_kernel[(n_blocks,)](
         packed_mask, block_counts, N=N, BYTE_BLOCK=triton.cdiv(block, 8),
     )
     block_prefix = torch.cat([
-        torch.zeros(1, device=vals.device, dtype=torch.int64),
-        torch.cumsum(block_counts, dim=0)[:-1].to(torch.int64),
+        torch.zeros(1, device=vals.device, dtype=torch.int32),
+        torch.cumsum(block_counts, dim=0)[:-1],
     ])
 
     out = torch.empty(N, device=vals.device, dtype=vals.dtype)
